Code/requirements.txt
absl-py==1.4.0
appdirs==1.4.4
asttokens==2.2.1
backcall==0.2.0
cachetools==5.2.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
distlib==0.3.6
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
fonttools==4.38.0
gitdb==4.0.10
GitPython==3.1.30
google-auth==2.16.0
google-auth-oauthlib==0.4.6
grpcio==1.51.1
idna==3.4
imageio==2.25.1
importlib-metadata==6.0.0
ipykernel==6.20.2
ipython==8.8.0
jedi==0.18.2
joblib==1.2.0
jupyter_client==7.4.9
jupyter_core==5.1.5
kiwisolver==1.4.4
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.6.2
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
networkx==3.0
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==23.0
pandas==1.5.2
parso==0.8.3
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.14.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.7.1
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.10.0
sentry-sdk==1.14.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
stack-data==0.6.2
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tifffile==2023.2.3
--extra-index https://download.pytorch.org/whl/cu116
torch==1.12.1+cu116
torch-tb-profiler==0.4.1
torchsummary==1.5.1
torchvision==0.13.1+cu116
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
typing_extensions==4.4.0
urllib3==1.26.14
virtualenv==20.17.1
wandb==0.13.9
wcwidth==0.2.6
Werkzeug==2.2.2
zipp==3.11.0
